refactor: 불필요한 인자 전달을 줄이도록 수정
