+++ src/main/kotlin/blackjack/domain/state/Blackjack.kt
@@ -5,8 +5,8 @@ import blackjack.domain.Hand
 import blackjack.domain.Result
 
 class Blackjack(override val hand: Hand) : Finished(hand) {
-    override fun decideResult(dealer: Dealer): Result {
-        return when (dealer.state) {
+    override fun decideResult(state: PlayingState): Result {
+        return when (state) {
             is Blackjack -> Result.PUSH
             else -> Result.WIN
         }
