+++ src/main/kotlin/blackjack/domain/state/Bust.kt
@@ -5,7 +5,7 @@ import blackjack.domain.Hand
 import blackjack.domain.Result
 
 class Bust(override val hand: Hand) : Finished(hand) {
-    override fun decideResult(dealer: Dealer): Result {
+    override fun decideResult(state: PlayingState): Result {
         return Result.LOSE
     }
 }
