+++ src/main/kotlin/blackjack/domain/state/Finished.kt
@@ -1,6 +1,5 @@
 package blackjack.domain.state
 
-import blackjack.domain.Dealer
 import blackjack.domain.Hand
 import blackjack.domain.Result
 import blackjack.domain.card.Card
@@ -10,7 +9,7 @@ abstract class Finished(override val hand: Hand) : PlayingState {
         throw IllegalStateException(ERROR_CANT_DRAW_CARD)
     }
 
-    abstract fun decideResult(dealer: Dealer): Result
+    abstract fun decideResult(state: PlayingState): Result
 
     companion object {
         private const val ERROR_CANT_DRAW_CARD = "더 이상 카드를 뽑을 수 없습니다."
