+++ src/main/kotlin/blackjack/domain/state/Stay.kt
@@ -5,19 +5,18 @@ import blackjack.domain.Hand
 import blackjack.domain.Result
 
 class Stay(override val hand: Hand) : Finished(hand) {
-    override fun decideResult(dealer: Dealer): Result {
-        return when {
-            dealer.isBust() -> Result.WIN
-            dealer.state is Bust -> Result.WIN
-            dealer.state is Blackjack -> Result.LOSE
-            else -> compareTo(dealer)
+    override fun decideResult(state: PlayingState): Result {
+        return when (state) {
+            is Bust -> Result.WIN
+            is Blackjack -> Result.LOSE
+            else -> compareTo(state)
         }
     }
 
-    private fun compareTo(dealer: Dealer): Result {
+    private fun compareTo(state: PlayingState): Result {
         return when {
-            this.hand.sum() > dealer.state.hand.sum() -> Result.WIN
-            this.hand.sum() < dealer.state.hand.sum() -> Result.LOSE
+            this.hand.sum() > state.hand.sum() -> Result.WIN
+            this.hand.sum() < state.hand.sum() -> Result.LOSE
             else -> Result.PUSH
         }
     }
