+++ src/test/kotlin/blackjack/domain/PlayingStateTest.kt
@@ -9,29 +9,5 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 class PlayingStateTest {
-    @Test
-    fun `딜러가 blackjack이고 플레이어가 bust이면 플레이어가 진다`() {
-        val dealer = Dealer()
-        val player = Player("name1", Money(1000))
-        dealer.state.hand.addCard(ClubKing)
-        dealer.state.hand.addCard(ClubAce)
-        val result = Bust(player.state.hand).decideResult(dealer)
-        assertThat(result).isEqualTo(Result.LOSE)
-    }
-
-    @Test
-    fun `딜러가 bust이고 플레이어가 bust이면 플레이어가 진다`() {
-        val dealer = Dealer()
-        val player = Player("name1", Money(1000))
-        val result = Bust(player.state.hand).decideResult(dealer)
-        assertThat(result).isEqualTo(Result.LOSE)
-    }
-
-//    @Test
-//    fun `딜러가 Stay이고 플레이어가 bust이면 플레이어가 진다`() {
-//        val dealer = Dealer
-//        val player = Player
-//        val result = Bust(player.state.hand).decideResult(dealer)
-//        assertThat(result).isEqualTo(Result.LOSE)
-//    }
+    //
 }
